requirements.txt
APScheduler==3.10.4
numpy==1.26.1
oss2==2.18.3
pandas==2.1.2
Pillow==10.1.0
pyquery==2.0.0
pytz==2023.3.post1
requests==2.31.0
Scrapy==2.11.0
pytest==7.4.3
inflection==0.5.1
openpyxl==3.1.2
